Fixed some disconnect / timeout issues.
Added smart_sleep after takeoff and land so that multiple commands are not sent.
Changed some controls to be (in MY opinion) more intuitive.
Fixed the yaw being reversed.
Made sure that the drone was landed before closing the pygame window with escape, made the drone disconnect before pygame window is closed.